pyproject.toml
[tool.poetry]
name = "babel-assistant"
version = "0.1.0"
description = ""
authors = ["margostino"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.1.5"
langchain-openai = "^0.0.5"
python-dotenv = "^1.0.1"
pydantic = "1.10.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
